feat: Add button and script to control lamp image
This commit introduces a button and associated JavaScript to toggle the lamp image between its default and active states. The button now has an `onclick` event handler that calls the `changeImage` function and an ID of "click" for easy access. A script tag has been added to load the custom script.